+++ packages/builtin_plugins/mflowy/builtin_plugins/loaders/parquet_loader.py
@@ -5,17 +5,17 @@ import logging
 from typing import Annotated
 
 import pandas as pd
-from mflowy.builtin_plugins.middlewares import log_load_profile
+from mflowy.builtin_plugins.middlewares import log_load_data_fingerprint, log_load_profile
 from mflowy.driver.handler import handler
-from mflowy.utils.file import read_bytes
 from mflowy.utils.path import ensure_relative_path_under_task_dir
 
 from . import report_loaded
+from .utils import set_data_fingerprint
 
 logger = logging.getLogger(__name__)
 
 
-@handler(log_load_profile)
+@handler(log_load_data_fingerprint, log_load_profile)
 def parquet(
     source: Annotated[str, "文件路径"],
     engine: Annotated[str, "Parquet 引擎"] = "pyarrow",
@@ -23,6 +23,7 @@ def parquet(
 ) -> pd.DataFrame:
     """从 Parquet 列式存储文件加载数据。"""
     path = ensure_relative_path_under_task_dir(source)
-    df = pd.read_parquet(io.BytesIO(read_bytes(path)), engine=engine, **kwargs)  # type:ignore
+    set_data_fingerprint(path.as_posix())
+    df = pd.read_parquet(io.BytesIO(path.read_bytes()), engine=engine, **kwargs)  # type:ignore
     report_loaded(df)
     return df